+++ src/data/doc-it/R16Roadmap.ts
@@ -1,28 +1,7 @@
-export type FrameworkFilter = 'all' | 'angular' | 'react' | 'ionic' | 'react-native';
-export type ItemTypeFilter = 'all' | 'design' | 'development';
-export type Status = 'backlog' | 'in-progress' | 'pre-release' | 'deferred' | 'finished';
-export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+import { Quarter, RoadmapBucket } from './_types';
 
 const CURRENT_YEAR = 2020;
 const CURRENT_QUARTER: Quarter = 'Q2';
-
-export type RoadmapItem = {
-    name: string;
-    description: string;
-    year: string | number;
-    quarter: Quarter;
-    status: Status;
-    author?: string;
-    type?: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-};
-type RoadmapBucket = {
-    name: string;
-    description: string;
-    type: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-    items: RoadmapItem[];
-};
 
 export const roadmap: RoadmapBucket[] = [
     {
@@ -176,12 +155,13 @@ export const roadmap: RoadmapBucket[] = [
             },
             {
                 name: 'Enhanced Internationalization Support',
-                description: 'Updating our Internationalization examples to support additional languages and use recommended tools',
+                description:
+                    'Updating our Internationalization examples to support additional languages and use recommended tools',
                 year: CURRENT_YEAR,
                 quarter: CURRENT_QUARTER,
                 status: 'finished',
                 author: 'PX Blue',
-            }
+            },
         ],
     },
     {
