+++ src/data/doc-it/R18Roadmap.ts
@@ -1,28 +1,7 @@
-export type FrameworkFilter = 'all' | 'angular' | 'react' | 'ionic' | 'react-native';
-export type ItemTypeFilter = 'all' | 'design' | 'development';
-export type Status = 'backlog' | 'in-progress' | 'pre-release' | 'deferred' | 'finished';
-export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+import { Quarter, RoadmapBucket } from './_types';
 
 const CURRENT_YEAR = 2020;
 const CURRENT_QUARTER: Quarter = 'Q3';
-
-export type RoadmapItem = {
-    name: string;
-    description: string;
-    year: string | number;
-    quarter: Quarter;
-    status: Status;
-    author?: string;
-    type?: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-};
-type RoadmapBucket = {
-    name: string;
-    description: string;
-    type: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-    items: RoadmapItem[];
-};
 
 export const roadmap: RoadmapBucket[] = [
     {
@@ -148,14 +127,14 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Reusable UI elements from @pxblue/react-components',
         type: 'development',
         framework: ['react'],
-        items: [ ],
+        items: [],
     },
     {
         name: 'React Native Component Library',
         description: 'Reusable UI elements from @pxblue/react-native-components',
         type: 'development',
         framework: ['react-native'],
-        items: [ ],
+        items: [],
     },
     {
         name: 'pxblue.github.io',
@@ -185,7 +164,7 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Basic code snippets showing how to implement various design patterns',
         type: 'development',
         framework: ['all'],
-        items: [ ],
+        items: [],
     },
     {
         name: 'Reusable Workflows',
