+++ src/data/doc-it/R19Roadmap.ts
@@ -1,35 +1,14 @@
-export type FrameworkFilter = 'all' | 'angular' | 'react' | 'ionic' | 'react-native';
-export type ItemTypeFilter = 'all' | 'design' | 'development';
-export type Status = 'backlog' | 'in-progress' | 'pre-release' | 'deferred' | 'finished';
-export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+import { Quarter, RoadmapBucket } from './_types';
 
 const CURRENT_YEAR = 2021;
 const CURRENT_QUARTER: Quarter = 'Q1';
-
-export type RoadmapItem = {
-    name: string;
-    description: string;
-    year: string | number;
-    quarter: Quarter;
-    status: Status;
-    author?: string;
-    type?: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-};
-type RoadmapBucket = {
-    name: string;
-    description: string;
-    type: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-    items: RoadmapItem[];
-};
 
 export const roadmap: RoadmapBucket[] = [
     {
         name: 'Figma Component Library',
         description: 'Updates to the components in the designer stickersheet',
         type: 'design',
-        items: [ ],
+        items: [],
     },
     {
         name: 'Figma Design Patterns',
@@ -107,7 +86,7 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Reusable UI elements from @pxblue/angular-components',
         type: 'development',
         framework: ['angular', 'ionic'],
-        items: [ ],
+        items: [],
     },
     {
         name: 'React Component Library',
