+++ src/data/doc-it/R20Roadmap.ts
@@ -1,28 +1,7 @@
-export type FrameworkFilter = 'all' | 'angular' | 'react' | 'ionic' | 'react-native';
-export type ItemTypeFilter = 'all' | 'design' | 'development';
-export type Status = 'backlog' | 'in-progress' | 'pre-release' | 'deferred' | 'finished';
-export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+import { Quarter, RoadmapBucket } from './_types';
 
 const CURRENT_YEAR = 2021;
 const CURRENT_QUARTER: Quarter = 'Q2';
-
-export type RoadmapItem = {
-    name: string;
-    description: string;
-    year: string | number;
-    quarter: Quarter;
-    status: Status;
-    author?: string;
-    type?: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-};
-type RoadmapBucket = {
-    name: string;
-    description: string;
-    type: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-    items: RoadmapItem[];
-};
 
 export const roadmap: RoadmapBucket[] = [
     {
