+++ src/data/doc-it/R21Roadmap.ts
@@ -1,35 +1,14 @@
-export type FrameworkFilter = 'all' | 'angular' | 'react' | 'ionic' | 'react-native';
-export type ItemTypeFilter = 'all' | 'design' | 'development';
-export type Status = 'backlog' | 'in-progress' | 'pre-release' | 'deferred' | 'finished';
-export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+import { Quarter, RoadmapBucket } from './_types';
 
 const CURRENT_YEAR = 2021;
 const CURRENT_QUARTER: Quarter = 'Q3';
-
-export type RoadmapItem = {
-    name: string;
-    description: string;
-    year: string | number;
-    quarter: Quarter;
-    status: Status;
-    author?: string;
-    type?: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-};
-type RoadmapBucket = {
-    name: string;
-    description: string;
-    type: ItemTypeFilter;
-    framework?: FrameworkFilter[];
-    items: RoadmapItem[];
-};
 
 export const roadmap: RoadmapBucket[] = [
     {
         name: 'Figma Component Library',
         description: 'Updates to the components in the designer sticker sheet',
         type: 'design',
-        items: [ 
+        items: [
             {
                 name: 'System Components',
                 description: 'Updated system components for keyboards, cursors, notifications, etc.',
@@ -44,7 +23,7 @@ export const roadmap: RoadmapBucket[] = [
         name: 'Figma Design Patterns',
         description: 'Adding design pattern examples to Figma',
         type: 'design',
-        items: [ 
+        items: [
             {
                 name: 'Visualization Components',
                 description: 'Design reusable visualization components',
@@ -60,7 +39,7 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Reusable UI elements from @pxblue/react-native-components',
         type: 'development',
         framework: ['react-native'],
-        items: [ 
+        items: [
             {
                 name: 'Themed RNP Components',
                 description: 'Improved theming support via pre-styled wrapper components for React Native Paper',
@@ -76,14 +55,14 @@ export const roadmap: RoadmapBucket[] = [
                 quarter: CURRENT_QUARTER,
                 status: 'finished',
                 author: 'PX Blue',
-            }
+            },
         ],
     },
     {
         name: 'pxblue.github.io',
         description: 'The PX Blue documentation site',
         type: 'design',
-        items: [ 
+        items: [
             {
                 name: 'Device / Network Commissioning',
                 description: 'New design patterns for setting up a hardware device',
@@ -99,7 +78,7 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Basic code snippets showing how to implement various design patterns',
         type: 'development',
         framework: ['all'],
-        items: [ 
+        items: [
             {
                 name: 'Updated Design Pattern Examples',
                 description: 'More comprehensive examples implementing the different elements of our design patterns',
@@ -115,7 +94,7 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Components for implementing common workflows across applications',
         type: 'development',
         framework: ['all'],
-        items: [ 
+        items: [
             {
                 name: 'Authentication Workflow',
                 description: 'Add support for username-based login in the authentication workflows ',
@@ -131,15 +110,15 @@ export const roadmap: RoadmapBucket[] = [
         description: 'Components for implementing common workflows across applications',
         type: 'all',
         framework: ['all'],
-        items: [ 
+        items: [
             {
                 name: 'Device & Channel Icons',
                 description: 'New icons to represent more electrical devices and channels',
                 year: CURRENT_YEAR,
                 quarter: CURRENT_QUARTER,
                 status: 'finished',
-                author: 'PX Blue', 
-            }
+                author: 'PX Blue',
+            },
         ],
     },
 ];
