+++ src/data/doc-it/_types.ts
@@ -1,0 +1,21 @@
+export type FrameworkFilter = 'all' | 'angular' | 'react' | 'ionic' | 'react-native';
+export type ItemTypeFilter = 'all' | 'design' | 'development';
+export type Status = 'backlog' | 'in-progress' | 'pre-release' | 'deferred' | 'finished';
+export type Quarter = 'Q1' | 'Q2' | 'Q3' | 'Q4';
+export type RoadmapItem = {
+    name: string;
+    description: string;
+    year: string | number;
+    quarter: Quarter;
+    status: Status;
+    author?: string;
+    type?: ItemTypeFilter;
+    framework?: FrameworkFilter[];
+};
+export type RoadmapBucket = {
+    name: string;
+    description: string;
+    type: ItemTypeFilter;
+    framework?: FrameworkFilter[];
+    items: RoadmapItem[];
+};
